Fix bazel (#561)
* Fixup bazel builds

* update some versions

+++ old_robots/y2013/Eve2013/src/main/java/com/gos/ultimate_ascent/commands/StartGyro.java
@@ -18,7 +18,7 @@ public class StartGyro extends GosCommandBase {
 
     @Override
     public void execute() {
-        m_angle -= m_chassis.getGyroAngle();
+        m_angle -= (int) m_chassis.getGyroAngle();
         m_chassis.setFieldAdjustment(m_angle);
     }
